+++ include/cli/cli.h
@@ -41,73 +41,13 @@
 #include "colorprofile.h"
 #include "history.h"
 #include "split.h"
+#include "historystorage.h"
+#include "localhistorystorage.h"
 
 #define CLI_DEPRECATED_API
 
-#include <fstream> // @@@
-
 namespace cli
 {
-
-    // @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
-
-    class HistoryStorage
-    {
-        public:
-            virtual ~HistoryStorage() = default;
-            virtual void Store(const std::vector<std::string>& commands) = 0;
-            virtual std::vector<std::string> Commands() const = 0;
-    };
-
-    class LocalHistoryStorage : public HistoryStorage
-    {
-        public:
-            explicit LocalHistoryStorage(std::size_t size = 1000) : maxSize(size) {}
-            void Store(const std::vector<std::string>& cmds) override
-            {
-                commands.insert(commands.end(), cmds.begin(), cmds.end());
-                if (commands.size() > maxSize)
-                    commands.erase(commands.begin(), commands.begin()+commands.size()-maxSize);
-            }
-            std::vector<std::string> Commands() const override
-            {
-                return std::vector<std::string>(commands.begin(), commands.end());
-            }
-        private:
-            const std::size_t maxSize;
-            std::deque<std::string> commands;
-    };
-
-    class FileHistoryStorage : public HistoryStorage
-    {
-        public:
-            explicit FileHistoryStorage(const std::string& _fileName = ".cli") :
-                fileName(_fileName)
-            {}
-            void Store(const std::vector<std::string>& cmds) override
-            {
-                std::ofstream f(fileName, std::ios_base::app | std::ios_base::out);
-                for (const auto& line: cmds)
-                    f << line << '\n';
-            }
-            std::vector<std::string> Commands() const override
-            {
-                std::vector<std::string> commands;
-                std::ifstream in(fileName);
-                if (in)
-                {
-                    std::string line;
-                    while (std::getline(in, line))
-                        commands.push_back(line);
-                }
-                return commands;
-            }
-        private:
-            const std::string fileName;
-    };
-
-    // @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
-
 
     // ********************************************************************
 
@@ -180,15 +120,30 @@ namespace cli
     public:
         Cli(
             std::unique_ptr<Menu>&& _rootMenu,
-            std::function< void(std::ostream&)> _exitAction = std::function< void(std::ostream&) >(),
-            std::unique_ptr<HistoryStorage> historyStorage = std::make_unique<LocalHistoryStorage>()
+            std::function< void(std::ostream&)> _exitAction = {},
+            std::unique_ptr<HistoryStorage>&& historyStorage = std::make_unique<LocalHistoryStorage>()
         ) :
             globalHistoryStorage(std::move(historyStorage)),
             rootMenu(std::move(_rootMenu)),
             exitAction(_exitAction)
         {
         }
-
+#if 0 // whiy this does not work?
+        Cli(std::unique_ptr<Menu> _rootMenu, std::unique_ptr<HistoryStorage> historyStorage) : 
+            Cli(std::move(rootMenu), {}, std::move(historyStorage))
+        {
+        }
+#else
+        Cli(
+            std::unique_ptr<Menu>&& _rootMenu,
+            std::unique_ptr<HistoryStorage>&& historyStorage
+        ) :
+            globalHistoryStorage(std::move(historyStorage)),
+            rootMenu(std::move(_rootMenu)),
+            exitAction(nullptr)
+        {
+        }
+#endif
         // disable value semantics
         Cli(const Cli&) = delete;
         Cli& operator = (const Cli&) = delete;